+++ exports/__init__.py
@@ -1,5 +1,5 @@
-"""Export model helpers."""
+"""Public exports package interface."""
 
-from .models import RecruitingWizardExport
+from .models import RecruitingWizardExport, apply_field_metadata_to_payload
 
-__all__ = ["RecruitingWizardExport"]
+__all__ = ["RecruitingWizardExport", "apply_field_metadata_to_payload"]